Node保存单独Id

+++ hw3d/App.cpp
@@ -107,7 +107,7 @@ void App::DoFrame()
 	{
 		if (!wnd.GetCursorEnabled())
 		{
-			camera.Rotate(delta->x, delta->y);
+			camera.Rotate((float)delta->x, (float)delta->y);
 		}
 	}
